rango_boom-backend/requirements/base.txt
Django
djangorestframework
django-cors-headers
django-filter
psycopg
channels
channels-redis
Pillow
celery
redis
django-treebeard